requirements.txt
Flask
Werkzeug
pymongo
python-dotenv
web3
py-solc-x